+++ mediachain/translation/utils.py
@@ -1,10 +1,3 @@
-import base58
-import json
-from pygments import highlight
-from pygments.lexers.data import JsonLexer
-from pygments.formatters.terminal import TerminalFormatter
-
-
 MEDIACHAIN_OBJECT_TAG = '__mediachain_object__'
 MEDIACHAIN_ASSET_TAG = '__mediachain_asset__'
 
@@ -26,19 +19,3 @@ def is_mediachain_object(value):
 def is_canonical(obj):
     return obj['type'] == 'artefact' or obj['type'] == 'entity'
 
-
-
-def stringify_refs(obj):
-    res = {}
-    for k, v in obj.iteritems():
-        if isinstance(v, dict):
-            v = stringify_refs(v)
-        if k == u'@link':
-            v = base58.b58encode(v)
-        res[k] = v
-    return res
-
-
-def dump(obj):
-    j = json.dumps(stringify_refs(obj), indent=2)
-    print highlight(j, JsonLexer(), TerminalFormatter())
